requirements.txt
# Core AI/LLM packages
openai>=1.3.8
langchain>=0.1.10
langchain-community>=0.1.0
langchain-text-splitters>=0.0.1
langchain-groq>=0.0.64


# Vector stores
pinecone-client>=2.2.1

# Web / API
fastapi>=0.111.2
uvicorn>=0.24.0
fastapi[all]
python-multipart>=0.0.6

# Utilities
pydantic>=2.4.2
requests>=2.31.0
python-dotenv>=1.0.0
asyncio

# Logging / debugging
python-json-logger>=2.0.7

langgraph>=0.0.64

#vision applications
open_clip_torch 
torch 
torchvision  
weaviate-client>=3.5.0
nltk
google-genai
gradio-client
pinecone-client
langchain_openai
langchain_google_genai